requirements.txt
anyio==3.3.4
appdirs==1.4.4
asgiref==3.4.1
astroid==2.4.2
attrs==21.2.0
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.3
colorama==0.4.4
colorlog==5.0.1
diff-match-patch==20200713
docopt==0.6.2
fastapi==0.70.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==2.10
isort==5.6.3
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lxml==4.6.3
mccabe==0.6.1
mysql-connector-python==8.0.27
online-judge-api-client==10.9.0
online-judge-tools==11.3.1
pipreqs==0.4.11
protobuf==3.19.1
pycodestyle==2.6.0
pydantic==1.8.2
pylint==2.6.0
PyMySQL==1.0.2
pyrsistent==0.17.3
python-dotenv==0.19.2
requests==2.25.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.5
uvicorn==0.15.0
wrapt==1.12.1
yarg==0.1.9